fix: properly show organizational unit details
Problem
The details view of a root administrative unit did not display any data. This behavior was caused by accesing the name property on a null organizational unit.

Solution
To use the optional chaining operator (?) with the nullish coalescing operator, so if the organizational's unit parent is null, because is a root ou, then show "N/A" string.