Dev (#18)
* fix: properly sort columns on activity logs datatable

* WIP fix: fixing activity log issues related

* feat: improve and translate activity log descriptions

Related to #2

* WIP fix: working on fixing bugs in the event details view

* WIP fix: working on fixing bugs in the activity log details view

* feat: improve the behavior of datatables

* refactor: reorganize folders and files

* feat: improve user management

Now users can be deactivated/activated. Deactivated users cannot login in the system due their deactivated status.

When activate an user if it was previosly deleted it will be automatically restored.

These statuses are shown in the index view of users.

Closes #13

* fix: properly show `vue-sooner` notification

Problem
Due to the upgrade on `vue-sonner` npm package the flash notifications wasn't showing properly.

Solution
Import the css file required for `vue-sonner`.